广和新闻-第一次修改

+++ src/main/java/com/guanghe/api/service/CompanyCultrueService.java
@@ -20,24 +20,39 @@ public class CompanyCultrueService {
     private CompanyCultrueDao companyCultrueDao;
 
     public List<CompanyCultrueBO> queryCompanyCultrueList( Map<String,Object> map){
+        if (map == null){
+            return null;
+        }
         List<CompanyCultrueBO> companyCultrue = companyCultrueDao.queryCompanyCultrueList(map);
         return companyCultrue;
 
     }
 
     public CompanyCultrueBO queryCompanyCultrueById(Integer newsId){
+        if (newsId == null){
+            return null;
+        }
         return companyCultrueDao.queryCompanyCultrueById(newsId);
     }
 
     public void deleteCompanyCultrue(Integer newsId){
+        if (newsId == null){
+            return ;
+        }
         companyCultrueDao.deleteCompanyCultrue(newsId);
     }
 
     public void addCompanyCultrue(CompanyCultrueBO news){
+        if (news == null){
+            return ;
+        }
         companyCultrueDao.addCompanyCultrueBO(news);
     }
 
     public void updateCompanyCultrueBO(CompanyCultrueBO news){
+        if (news == null){
+            return ;
+        }
         companyCultrueDao.updateCompanyCultrueBO(news);
     }
 }
