+++ src/main/java/com/guanghe/api/service/IndustryInformationService.java
@@ -20,23 +20,39 @@ public class IndustryInformationService {
     private IndustryInformationDao industryInformationDao;
 
     public List<IndustryInformationBO> queryIndustryInformationList( Map<String,Object> map){
+        if(map == null){
+            return null;
+        }
         List<IndustryInformationBO> industryInformation = industryInformationDao.queryIndustryInformationList(map);
         return industryInformation;
     }
 
     public IndustryInformationBO queryIndustryInformationById(Integer newsId){
+        if (newsId == null){
+            return null;
+        }
+
         return industryInformationDao.queryIndustryInformationById(newsId);
     }
 
     public void deleteIndustryInformation(Integer newsId){
+        if (newsId == null){
+            return ;
+        }
         industryInformationDao.deleteIndustryInformation(newsId);
     }
 
     public void addIndustryInformation(IndustryInformationBO news){
+        if (news == null){
+            return ;
+        }
         industryInformationDao.addIndustryInformationBO(news);
     }
 
     public void updateIndustryInformationBO(IndustryInformationBO news){
+        if (news == null){
+            return ;
+        }
         industryInformationDao.updateIndustryInformationBO(news);
     }
 }
